Cythonize _update_dict function
Lower the `_update_dict` into Cython (as close to C as possible). As the
big performance pieces are in the BLAS and not in this code, the main
objective here is to cutdown the amount of time not spent in those BLAS
functions as it can cause the BLAS to spin down its threadpool for
instance only to spin it up fractions of seconds later, which is quite
costly. This avoids that by calling directly into the BLAS C API with
contiguous arrays back to back. Only little bits of Cython are used to
string these together.