Patch out SYRK portions in GEMM
The GEMM functions in ATLAS's reference BLAS will detect if a matrix is
being multiplied by its transpose and if so dispatch to SYRK. While a
good idea, it is much more complicated to pull in the SYRK portions of
ATLAS as they have some headers and source files that are generated
dynamically at build time. Given our changes here do not benefit from
SYRK, simply patch out those custom dispatch portions in the GEMM
functions.